organized c and h file, question: do inline functions bodies belong in header file
